add templated unit test
