+++ test/core/CMakeLists.txt
@@ -15,10 +15,10 @@ neofoam_unit_test(parallelAlgorithms)
 
 # The registerClass test has to be created without linking against neofoam. Otherwise, it will pick
 # up the factories registered in neofoam, which will make the test hard to maintain.
-add_executable(registerClass "registerClass.cpp")
-target_link_libraries(registerClass PRIVATE Catch2::Catch2WithMain cpptrace::cpptrace
-                                            NeoFOAM_public_api)
+add_executable(runTimeSelectionFactory "runTimeSelectionFactory.cpp")
+target_link_libraries(runTimeSelectionFactory PRIVATE Catch2::Catch2WithMain cpptrace::cpptrace
+                                                      NeoFOAM_public_api)
 add_test(
-  NAME registerClass
-  COMMAND registerClass
+  NAME runTimeSelectionFactory
+  COMMAND runTimeSelectionFactory
   WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
